membind: FIRSTTOUCH doesn't want a partial nodeset (except on AIX)
Linux already ignored nodesets passed for FIRSTTOUCH.
Now return an error when not passing an entire set,
just like we do for HP-UX and Solaris.

Document that FIRSTTOUCH should usually get a full nodeset,
except on AIX where it can fallback to the given non-local nodes
specified in the nodeset.

Closes #237.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>