linux/membind: don't use MPOL_DEFAULT for FIRSTTOUCH
MPOL_DEFAULT falls back to the larger policy (area, then thread, then system-wide).
If we set FIRSTTOUCH on an area but the thread has another policy,
the area's MPOL_DEFAULT will cause the thread policy to be used.

Linux has MPOL_LOCAL for actual FIRSTTOUCH instead of MPOL_DEFAULT,
but it's only available since 3.8.
It's implemented (and returned) as PREFERRED with no nodeset in
the kernel, which has been supported for a long time, so just use this.

Refs #237

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>