windows: replace get_area_membind() with get_area_memlocation()
QueryWorkingSetEx() returns either the current node of the page,
or nothing if not allocated (we must check the Valid field before
reading the Node field).

Windows uses first touch for not-bound buffers
(those allocated with VirtualAlloc(), not with VirtualAllocExNuma()),
and falls back to other nodes otherwise.
So there's really no way to find-out area membind with QueryWorkingSetEx().

Fortunately, get_area_membind() was disabled by mistake since 1.11.1
so that buggy behavior wasn't enabled recently.
So we're pretty much just enabling a new feature here.

Closes #284.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>