doc: update/fix PPC64 images
We modified source XMLs in the previous commits.
Now fix the way to generate images, document it in doc/images/HACKING,
and actually commit those final images.

We must use logical index (to match what's said in the text).
Slightly improve that text.

Also force the horizontal layout and reduce the width of the
smtoff images in the PDF
(so that these 3 images have the same scale and comparable layouts).

Also remove the cache indexes for clarity.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>