membind: error-out with EINVAL when invalid policies are given
Check that the policy is correct, in case somebody passes
a flag in the policy instead of inside flags.

Backends only returned ENOSYS (and sometimes ignored the binding error
in non-strict mode), we better error-out for such buggy values.

In theory, this breaks the ABI in case somebody ever passed buggy values
without getting an error, but I am going to ignore this issue.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>