pci: serialize pciaccess-based discovery with a mutex
libpciaccess uses an internal global variable to describe the PCI
hierarchy without internal refcounting. If multiple threads build
a topology simultaneously, the first one will destroy the pciaccess
stuff that others still need.

Moreover, there are internal arrays that are dynamically reallocated
and populated with vendor/device names according to caller's requests.

Put a mutex around all of this discovery. In the mono-threaded case,
the overhead of a single lock/unlock is negligible. In the
multi-threaded case, pci discovery for individual topologies are
serialized, which isn't very good, but people are encouraged not
to build multiple topologies in multiple threads anyway.

Closes #297

Finer grain serializing would require refcounting of users and
careful review of pciaccess internal behavior with respect
to dynamic realloc/populate of at least device/vendor names.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>