lstopo/ascii: ignore --fontsize and --gridsize
We cannot actually change the fontsize in text mode,
while changing the gridsize results in scaling everything-but-text backwards.

Terminals are usually too small for modern platforms anyway,
so don't bother letting --gridsize change the room around the text.

By the way, cleanup the code for scaling between generic lstopo-draw coordinates
and text coordinates.

This reverts the previous commit which was only useful for backporting
to stable branches.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>