x86: use the number of existing CPUs instead of only online ones
On a 4-core machine with #2 offline, we want x86 to look at all 4 cores,
so that we get info for all online ones (#1, #3 and #4).
Before this commit, x86 would only look at the 3 first ones
(just like if the offline PUs were always at the end of the range).

At least Linux and Solaris can put some CPUs offline in the middle of the index range.
So ask hwloc_fallback_nbprocessors() to get the number of existing CPUs.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>