VERSION: bump soname from 12:0:0 to 15:0:0
hwloc 2.0.0 was erroneously set to 12:0:0 while 1.11.9 was 12:6:7.

Bump it in 2.0.1 so that we avoid conflicts between 1.x and 2.x libs.
Bump it to 15:0:0 instead of 13:0:0 in case we ever have to add something
to the v1.x API.

This means that existing applications will need to be recompiled against 2.0.1.
It shouldn't too bad since 2.0.0 is recent and wasn't much deployed yet,
so better fix the ugliness now that later.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>