lstopo: display both physical and logical indexes by default
Finally drop the confusion between lstopo console and graphical modes
which used to display different indexes by default.

In graphical, we display both by default for PU and NUMA, none for other objects.
In textual, we display all logical indexes, and only physical for PU and NUMA.

-l and -p switch to only logical or only physical for all objects,
both in console and in graphical mode.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>