core: document that get_allowed_resources() may be called twice during discovery
Linux/BGQ/Solaris define that hook, use it internally during their own discover callback(),
and then let the core call it again if THISSYSTEM_ALLOWED_RESOURCES is set in flags
or in the environment.

THISSYSTEM_ALLOWED_RESOURCES shouldn't be set when loading from the native backend,
so it's unlikely that this duplicate call is visible in practice.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>