manpages: don't use # for manual substitution
AIX7 make fails to generate manpages since hwloc 1.11.4 (commit 20074f998)
because the end of the SEDMAN variable, once substitued with '\#...',
is interpreted as a comment.

gmake works fine and nobody has complained for 2 years, good to know.

Don't use #FOO# but rather £FOO£ for variables that we manually
substitute with SEDMAN in manpages. No need to escape that character,
and no risk of having it interpreted by make.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>