linux/knl: complete rework of heuristics to guess cluster/memory mode
Using empty cpusets isn't reliable because some CPUs may be offline.
Using memory size isn't reliable because part of the memory may be offline.
mOS uses both, causing failures.

First do some stats on the NUMA distance matrix, whose layout is always
the same, except for the ordering of MCDRAMs. Values in that matrix are
always the same as far as I know (10/21/31/41), but it is not guaranteed,
so don't hardwire them.

Use the number of nodes and the matrix layout to guess the modes when
hwloc-dump-hwdata didn't give them.
Quadrant/SNC4 are the only recommended cluster modes, so use them
unless we're sure it's rather SNC-2.
When we're not sure about Flat vs Hybrid, assume Flat.

Add tons of checks and abort the KNL quirk whenever something unexpected
occurs.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>