restrict: reorder children when modified
If we first bit of the cpuset of some children increases during restrict(),
these children might become out-of-order.

Example:
  Core P#0
    PU P#0
    PU P#2
  Core P#1
    PU P#1
    PU P#3
If restrict() removes PU P#0, Cores must be swapped because of
their updated cpusets.

In v2.x, reordering might also be needed for memory children
but it's so unlikely and harmless that we don't do it yet.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>