windows: add a test to verify that MSVC version/soname match the official numbers
MSVC hardwires the VERSION numbers/strings and soname in its manually generated
*_config.h files and in the vcxproj files. Make sure they match what we have
in normal headers and VERSION files.

Only check the soname when in a release branch (ignore on master, where the
lib version is 0:0:0).

This test doesn't run automatically in make check because it requires bash
and grep supporting -P. There's no need to run it for each build on each slave
anyway. So just run it manually in the main jenkins job.

Signed-off-by: Valentin Hoyet <Valentin.Hoyet@inria.fr>
Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>