core: make sure that cousin pointers are sane after merging root with an identical child
When merging identical parent/child, the root may be replaced with a child.
Make sure we clear old cousin pointers in that child.
Usually they are cleared when rebuilding levels, but the root level is never rebuilt.

This fixes a debug assert on a dual-PU Solaris machines where the second PU is offline.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>