lstopo/draw: add a native SVG backend
Generate the SVG as plain text with id and class attributes so that
we may identify and manipulate SVG objects later. This is useful
for incorporating the SVG in a webpage and having HTML/JS play
with it.

The output is slightly less pretty than Cairo because we don't know
what the textbox size should be. Hence we use a monospace font to
simplify things.

This backend is always available, using the "nativesvg" output format.
When Cairo is built-in, is is used by default.
When Cairo isn't available, nativesvg is used.

Fixes #267

Thanks to Sebastien Valat for the idea and the draft.
See https://github.com/svalat/htopml for an example of use.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>
Signed-off-by: Valentin Hoyet <Valentin.Hoyet@inria.fr>