lstopo/draw: fix the displaying of multiple NUMA nodes as normal children order
When "--children-order plain" was specified, all children (including memory)
are placed by the main call to place__children() but it wasn't called if
only memory_arity was > 0. That caused all memory children to be placed
as the first child on the left.

Only matter when there are >1 NUMA nodes in the same parent.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>
(cherry picked from commit ae7999c5694ec6355950d6fbf455a02b056ea7fb)