linux: add a debug env var to allow nodes with overlapping cpusets
Linux and current BIOS/SRAT enforces that different NUMA nodes cannot
be close to the same PUs according to /sys/devices/system/node/node*/cpumap.
We ignore them if that ever happens (we've seen such buggy BIOS in the past).

Don't ignore them if HWLOC_DEBUG_ALLOW_OVERLAPPING_NODE_CPUSETS=1
is set in the environment because it's useful for regression-testing
the core with complex NUMA node hierarchies.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>