API: add a new MEMCACHE object type for Memory-side Caches
These caches are coming in Intel platforms, and they will be exposed
by the ACPI HMAT table and Linux 5.2+.
KNL MCDRAM in Cache was actually a memory-side cache but the platform
was not ready to expose it.

This is a new (second) "memory" object type (the other one is NUMA node).
It's not considered a "cache" type as other usual CPU-side caches.

Object attributes are identical to usual cache attributes.
These caches have their own virtual depth.
They will be placed between CPU-side parent and NUMA nodes,
in the memory hierarchy of children.

e.g. on KNL in Hybrid mode:
Group(Cluster)
  MemCache
    NUMANode(DDR)
  NUMANode(MCDRAM)
  L2
   PU
   ...

Memory-side caches are disabled by default to avoid disturbing existing
applications where we usually want NUMA nodes as immediate memory children
of CPU-side objects.

This breaks the shmem ABI, which means 2.1 won't be able to share topologies
in shmem with 2.0.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>