tools: update for memory-side caches
lstopo/graphics:
When there's a memcache, we always put memory children inside a dedicated
memory box. The cache is placed above its NUMA node(s), without a vertical
space between, so that we know which NUMA node is cached.

all:
--no-caches or --filter cache:... applies to both cpu-side and memory-side
caches since both are enabled by default in tools (contrary to the API).

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>