lstopo/draw: add LSTOPO_TEXT_XSCALE environment variable
Backends that can't find the exact width of a text box (fig and nativesvg)
use a heuristics to do so. We over-exagerate the max possible width based
on wide letters.

This environment variable is a float number that scales that guessed width
in case we ever guess something too low (or way too large).

The scaling applies to all text boxes.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>