linux/sysfsnode: ignore NVIDIA GPU memory NUMA nodes
On P9, V100 GPU integrated memory is exposed as an additional NUMA node in Linux.
Hence it's possible to allocate memory there, but NVIDIA discourages this in
favor of using the NVIDIA/CUDA APIs. Also interleaving memory on all nodes
ends up allocating on GPUs too.

Get the association between GPUs and NUMA nodes from
  /proc/driver/nvidia/gpus/<pcibusid>/numa_status
and ignore those NUMA nodes.

Thanks to Max Katz and Edgar Leon for the help.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>