linux: fix management of openat() and fsroot
We would always fail if openat() was not supported or if open("/") failed
because hwloc_checkat() enforces root_fd != -1.
Fortunately, openat() is supported since 2.6.16 (when sysfs topology was added).
However, open("/") fails on Android 7 because / is --x.

fooat() functions actually support -1 as at dirfd, but the given path
has to be absolute.
Hence, in the normal case (real fsroot), no need to open("/") anymore,
but need to keep starting slashes in paths.
In the modified-fsroot case, we still open that fsroot and remove starting slashes
to make the path relative.

It means we don't have one fd open per alive topology anymore in normal case.
Fixes #115

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>