linux/sysfsnode: use sysfs HMAT initiator information to improve NUMA locality
Starting in Linux 5.2, sysfs node directories may contain information
about memory target/initiators. Use this to detect the actual locality
of NUMA nodes that are otherwise reported as CPU-less.

This is what KNL couldn't do (because each CPU may only be local to a
single NUMA node in Linux and in the SRAT ACPI table), but the ACPI HMAT
table (ACPI spec >= 6.2) solves this.

This will be useful for machines with two kinds of memory (DDR+PMEM, DDR+HBM)
where we expect to see DDR as default node (non-CPU-less) and other memory
(CPU-less with HMAT information).

We still insert default nodes first (those with non-empty cpumap in sysfs,
likely DDR) because we want these nodes to appear first in the list of local nodes.

However, other nodes may also be attached somewhere else (e.g. DDR at SNC level
and PMEM at Package level) hence we'll need an API to help choose the target node.

The ACPI HMAT table may also contain performance information, which is already
exposed in sysfs in Linux 5.2 but not used in hwloc yet.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>