lstopo/draw: add factorizing of identical children
Whenever there are more than 4 objects, show the first, second and last ones,
and replace others with dots.

This is enabled by default, but may be configured with --no-factorize
and --factorize, either globally or on a per-type basis:
one may change the number of objects before factorizing gets enabled,
and how many first and last children to keep out of the dots.

The 'f' shortcut now toggles factorizing and PCI collapsing.
'c' doesn't toggle collapsing anymore.

Factorizing only applies to normal CPU-side children.
It is similar to PCI collapsing, but the latter is pretty much always
enabled, hence we don't want to always toggle them simultaneously.

Closes #228

Signed-off-by: Valentin Hoyet <Valentin.Hoyet@inria.fr>
Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>