core: use phases for component types and for splitting discovery steps
Replace the single component type with a OR'ed mask of "phases".

For each phase, the discover() callback of each relevant backend is
invoked with the phase as an argument.

Phases may be excluded individually, so that a backend is either
entirely or partially disabled.

For now, we still don't have any component with multiple phases,
they are still GLOBAL, CPU or MISC for now.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>