xml/distances: add support for heterogeneous distances
Use a new "distances2hetero" tag instead of adding hetero-specific attributes
to the existing "distances2" tag. The latter would fail to load entirely in pre-2.1
releases, while the former just ignores the remaining of the XML (not an issue
since the new tag is always after everything supported in pre-2.1).

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>