core: add a name to internal distances structure
Not exported to users yet. Useful for better describing distances structure,
especially for upcoming matrices for objects with different types.

Only "NUMALatency" is currently used, and that's not very useful since
the distances structure type and kinds provide the same info. But it will
be more useful for identifying specific matrices such as the upcoming NVLink one.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>