keep_structure: only merge levels which are actually parent/children
We never checked the parent/child relationship because the number
of object in (unrelated) consecutive levels (in depth order) is
very often different.
However this is really needed in practice.

Fixes #363

Thanks to Mark Allen for reporting the issue.
He had a P9 machine restricted to only 6 cores.
And there were also 6 groups for CPU-less GPU memory nodes.
Those levels are really unrelated, but the core algorithm
decided to place them nearby in the list of depths.

In git master, this issue would only occur if we disable
the auto-ignoring of these GPU nodes.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>