utils: fix --flags/--thissystem when using a fake fsroot or cpuid directory
fsroot and cpuid are implemented in tools using environment variables
(those debug cases are not in the API since v2).
Those backends forced by environment variable override the normal
topology thissystem flag that may be set with set_flags() in the API
and with --flags or --thissystem in cli tools. One must use the
HWLOC_THISSYSTEM envvar to force the this system flag.
Implement this automatically in the tools (common helpers).

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>