linux: add support for cgroup v2
v2 cgroups use a unified hierarchy for all controllers,
where a parent can restrict to something different than a child.
We must use the "effective" cpus and mems masks to get the
actual intersection of all restrictions.

v1 and v2 cgroups may coexist, we check the v2 controllers
to see if "cpuset" is enabled there, or look for v1 with a cpuset
options, or for the old "cpuset" alias.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>