opencl/nvidia: properly detect the PCI domain of NVIDIA GPUs
Instead of hardwiring it to 0.

NVIDIA finally implemented this in CUDA 10.2. I couldn't find it
in their OpenCL extenstions, but they told me it's implemented.
It's the very next extension value (0x400a), it reports the right
result, and following extension values (0x400b+) still report errors.

Thanks to François Rue and the wonderful PlaFRIM team for the help.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>