doxy: disable doc building when using doxygen 1.8.16 inside the source tree
When we have the "images" subdirectory twice in doxygen's IMAGE_PATH
(one in @srcdir@ from prebuilt images, and one for @builddir@ for rebuilt images).
When srcdir==builddir, doxygen 1.8.16 warns that images are
"ambiguous" (because each image is available from both paths)
and just ignores them, breaking the PDF build.

This will be fixed in the next doxygen release thanks to commit
https://github.com/doxygen/doxygen/commit/752753c044af46cc4a22f83be8e4e1ae6be8f373

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>