hwloc-calc: improve logical/os index conversion of NUMA nodes
-I and -N must look at nodesets rather than cpusets when reporting NUMA nodes.
Enable this when --nodeset(-output) is given.

This is useful on machines with NUMA nodes at different levels
(e.g. DDR at SNC level and NVDIMM at socket level).
$ hwloc-calc --nodeset node:X -I node
now only reports X instead of also reporting the local node of the other kind of memory.

Hence now we may convert NUMA node logical to physical indexes with:
$ hwloc-calc --nodeset --physical-output node:2 -I node
3
and the reverse conversion:
$ hwloc-calc --nodeset --physical-input node:3 -I node
2

Thanks to Andres Rubio for the report.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>