pci: fix the parsing of osdev pci domain
The 'tmp' variable used to point to the previous '/' in the path,
causing the 32bits domain code to break domain parsing
by ignoring the first digit in the first [domain:]bus:device.function
subdirectory of the device path.
Usually 16bits domain started with 0 so it didn't break any existing test.

Change the 'tmp' variable to always point to the BDF instead of the
previous '/'.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>