nvml: add a NVLinkBandwidth matrix
Objects are GPU "nvml" OS devices for GPUs,
Package for POWER OpenCAPI/NVLink CPU ports,
and a PCI device for NVSwitch.

We assume that NVLink peers of PCI class 0x0680 and vendor IBM
are NVLink ports on the CPU.
That's true for POWER8/9 but no reply from IBM about the
future or some sort of standardization of these OpenCAPI ports.

We also assume that NVLink peers of PCI class 0x0680 and
vendor NVIDIA are NVSwitches.
NVIDIA doesn't provide any way to distinguish peer types.

POWER8 doesn't have correct PCI locality for these OpenCAPI
ports, hence we use the GPU locality instead,
unless HWLOC_NVML_USE_OPENCAPI_LOCALITY is set in the environment.

We will need some high level functions to summarize this matrix
because it's not obvious on DGX platforms with 16 GPUs and a NVSwitch:
* there are two independant set of connected GPUs/Switch,
  inter-switch links are not reported
* there are 6 hardware switches on each side, connected to one of
  the 6 links of the gpus

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>