utils/hwloc-bind: cleanup topology options management
Enforce those options (--restrict/--disallowed/--no-smt/--no-hbm)
at the beginning of the command-line instead of reupdating the
topology later if options and locations are mixed.

Try to document this but we don't show a precise error message
if --restrict/... is given too late.

In theory smt/hbm options could be parsed later but it seemed
more clear this way.

Signed-off-by: Valentin Hoyet <valentin.hoyet@inria.fr>
Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>