Add a "levelzero" backend
"Level Zero" is the lowlevel API for managing oneAPI devices.
We create "ze%d" devices for these GPUs (mostly Intel for now),
because "ze" is the prefix for function names in the levelzero API.

CPPFLAGS should point to ze_api.h that is currently installed
under $prefix/include/level_zero by the oneAPI compute runtime.

This code is for the L0 1.0 API, it won't work with previous
releases (e.g. 0.91).

We try to enable the required "Sysman" L0 sub-API but it may be
too late. More to come about this.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>