ci.inria.fr: allow to disable checks when launching jobs manually
When NO_CHECK is set in the jenkins UI for launching the job:
* make distcheck is replaced by make dist
* make check is disabled
* renaming check is also disabled since renaming is often done
  once new APIs are stabilized

This is useful during development, when testing portability, adding
printf for debugging backends, etc... anything that could break tests
(because of spurious lines in the output) but should not prevent
other pipeline steps from running.

Also it makes the CI several minutes faster, especially the extended
job on Windows.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>