lstopo: Add output option for LaTeX tikzpicture
This output option generates a standalone picture to be compiled with
pdflatex.  The colors are defined using the xcolor package.  If only the
tikzpicture is of interest, one can extract the image code by copying
the lines between
"%%%%% If inserting in another document, this is the actual source code of the picture %%%%%"
and
"%%%%% End of actual source code of the picture to insert in another document %%%%%",
however, some lines also have to be copied to the document preambule.

Signed-off-by: Clement Foyer <cfoyer@hpe.com>