API: add hwloc_get_local_numanode_objs()
This function returns an array of "local" NUMA nodes.
By default, "local" means matching the exact given locality.
Flags may extend that to larger or smaller locality, or to all nodes.

This function isn't specific to memattrs, but it will be the main
way to select nodes before looking at their attribute values.
Also it uses the new "localion" structure which is still only in memattrs.h
for now.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>