API: add HWLOC_TOPOLOGY_FLAG_IMPORT_SUPPORT to import support bits from XML
Useful when you want to check whether binding/... is supported on the distant
platform. OMPI currently passes this info on the side of the XML.

The support bits are still the local ones by default, but may be
overwritten with the XML ones when this topology flag is set.
The flag doesn't change the fact that binding is disabled by default
when loading from XML.

Most tools enable this flag so that imported support bits are propagated
in output XMLs.

Export support to XML may be disabled for debugging with HWLOC_XML_SUPPORT_EXPORT=0.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>