Look for libtinfo too when detecting ncurses.
When ncurses is compiled with --with-termlib, the tparm function lives
in libtinfo. Some distro's seem to provide both libncurses and libtinfo,
where libncurses is just a linker script to actually link against
libtinfo:

$ docker run ubuntu:18.04 /bin/bash -c '\
    apt-get update -qq &> /dev/null && \
    apt-get install -qq libncurses5-dev &> /dev/null \
    && cat /usr/lib/x86_64-linux-gnu/libcurses.so'
INPUT(libncurses.so.5 -ltinfo)

but it is better to look for tinfo ourselves in case this is missing.

For reference, see LLVM's detection of ncurses:

https://github.com/llvm/llvm-project/blob/3812919dc0588eeb74d440d3e3d8b1a097756f95/llvm/cmake/config-ix.cmake#L119

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>