x86: fix cache ids on Intel
Use the official way to compute the cache id on Intel
(same one as in the Linux kernel) by rounding-up
nbthreads_sharing and masking lower bits of the APIC ids
(instead of dividing the APIC id by nbthreads_sharing).

This should not affect any existing platform topology,
but it's more future-proof.

Thanks to Jonathan Peyton for the help.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>