API: add hwloc_topology_refresh()
Refreshes internal caches (distances and memattrs for now)
after modifying the topology (restrict, etc) and before
consulting the topology in a multithread context.

This is convenience function to avoid dummy calls to
hwloc_distances_get() or hwloc_memattr_get_value()
as explained in the Thread safety section in the doxy file.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>