core: don't report an big OS error when the insert comes from the user
If the "reason" for inserting an object is NULL, dont report
a big user-visible OS error warning. Just let the caller warn
when it gets an error code and if it wants to warn.
That's what lstopo --ps does when it fails to insert a group
for attaching a Misc for showing task binding.

Valentin properly handled this case during the callback->reason
conversion but I removed it mistake when I pushed all this as
commit 0dce1af93d1abf701817d0593099d484f13859ca

Thanks to Philippe Swartvagher for the report.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>