windows: use RtlGetVersion() for reliable version checks without manifest
GetVersion() works fine inside hwloc now that we build with a manifest.
However, external users will need a manifest too when building their
program so that their hwloc enables Windows10+ code properly (even if
they use a libhwloc that was built with a manifest). Not easy.

Use the RtlGetversion() symbol to get the real Windows version.
This is sort of ugly but many people have used this for a while,
hence Microsoft will likely have a hard time breaking it.
If the symbol is ever missing, we'll fallback to manifest+GetVersion.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>