core: fix inserting of custom groups replacing an existing ones
When inserting a group identical to an existing one,
the new one may "replace" the old one (depends on their kinds).
If so, we need to rebuild the level, fill its cpusets,
recompute its depth and total_memory, etc.

This happened when lstopo --top inserted a group (kind 0 for now)
identical to a NUMA group (kind higher).

Thanks to Scott Hicks for reporting the issue and providing
lots of debugging information.

Closes #446

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>