API: add hwloc/windows.h for querying windows processor groups
Windows processor groups are not a physical hierarchical level
(although it may be exposed by hwloc if it's different from
 NUMA nodes, packages, etc) but it has an impact of application
because binding must remain inside a single group.
Expose those groups explicitly so that applications know what
their binding masks should be restricted to.

Thanks to Ivan Kochin for the suggestion.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>