hwloc-ps: don't apply --pid-cmd to processes that will be filtered out
Split the calling of the user-provided command out of the main
common-ps functions and let the caller calls it after filtering
interesting processes.

Otherwise hwloc-ps --pid-cmd takes several seconds on all platforms
since we have to query hundreds of (uninteresting because not bound)
processes.

Thanks to Valentin Hoyet for the report.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>