API: add HWLOC_TOPOLOGY_FLAG_RESTRICT_TO_{CPU,MEM}BINDING
These flags apply a restrict during hwloc_topology_load()
and directly ignore excluded PUs or NUMAs.
The main goal is to avoid rebinding the x86 discovery to
all PUs since it breaks the process affinity on Windows
(see #366).

The THISSYSTEM flag is required as well since binding
is needed.

It only matters for PU but people may want to restrict
to NUMA for consistency, hence we provide both flags even
if NUMA discovery doesn't break Windows binding (and CPU
and memory binding are basically identical on Windows).

The implementation could be improved to filter excluded
PUs/NUMAs early in the backend, but that would require
lots of changes and testing.
For now, we only update the x86 backend to avoid breaking
the binding, and the core filters things out at the end
of the discovery.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>