Make -ltermcap the last option for ncurses
On macOS there seems to be a symlink from /usr/lib/libtermcap.dylib to
libncurses.5.4.dylib, which is the first option hwloc currently
considers for detecting ncurses. If you install ncurses with brew or
spack, you don't get this symlink, so hwloc will always link to the
system libncurses instead of the brew/spack version.

By considering libncurses first this should be fixed.

Signed-off-by: Harmen Stoppels <me@harmenstoppels.nl>