linux: fix support for NUMA node0 being offline
Just like we didn't support offline CPU#0 until commit
7bcc273efd50536961ba16d474efca4ae163229b, we need to
support node0 being offline as well.
It's not clear whether it's a new Linux feature or not,
this was reported on a POWER LPAR VM.

The symptoms are different here because we got no NUMA
nodes at all, hence the core hwloc added a default
machine-wide node. But this node got marked disallowed
by Linux cgroups. Hence load() failed with
 "Topology does not contain any NUMA node, aborting!"

We opportunistically assume node0 is online to avoid
the overhead in the vast majority of cases. If node0
is missing, we parse "online" to find the first node.

Thanks to Jirka Hladky for the report.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>