lstopo/x11: fix autoresizing when window is maximized
If the window is maximized, pressing + or - to change the scale
raises a ConfigureNotify event (as if the window had been resized)
that causes the scaling to go back to fit the window.

Detect whether the window is maximized during each ConfigureNotify
and don't ask for a scale update when we're maximized.

If the window is maximized in a single dimension, keep the current
behavior because zoom in/out will unmaximize it.

Thanks to Clement Foyer for the report.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>