utils/gather-cpuid: bring back pconfig
It's not clear what happened to pconfig in the x86 manual.
The doc was clear when first added. Now most of it moved to
a specific section without keeping all details about the
CPUID leaf. One may read this as if the subleaf value is the
entire EAX but everybody, including Linux, still uses bits 0:11
as in the old documentation.

So just revert 8e42ed34984ea36c2f6cd1dacd891e2865bdfaa8
and assume the (new) doc needs to be clarified.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>