xml/export/v1: fix distances export vs v1 memory group insert
When exporting some memory children to v1, we have to insert intermediate
Groups in some cases. Those groups did not have their parent pointer
initialized because it wasn't needed... except when deciding whether
a distance matrix should be exported or not: only machine-wide matrices
are exported (matrices attached to an object whose parent is NULL).

This caused distance matrices to be exported multiple times (in the root
and in some subgroups) and warnings on import (duplicates were ignored).

Thanks to Ralph Castain for the report.

Ref #494.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>