linux: add "Cluster" groups for Linux 5.16+ cluster topology
Expose "Cluster" group objects on Linux kernel 5.16+ for CPUs
that share some internal cache or bus. This can be equivalent
to the L2 Cache level on some platforms (e.g. x86) or a specific
level between L2 and L3 on others (e.g. ARM Kungpeng 920).

Based on a patch by Jonathan Cameron <Jonathan.Cameron@huawei.com>

These allow us to represent groups of cores where they are grouped
because of shared elements other than caches.  Typical example
would be a heirarchical coherency fabric with a cluster sharing a
single bridge to a wider coherency network on chip.

In ACPI this is represented in PPTT as a processor hierarchy node
structure above the level representing a CPU (possibly with multiple
threads) and below the hierarchy node structure that represents
either a die, or the level at which the last level cache is shared.
An example of this is the Kunpeng920 SoC which has 4 CPU cores
per cluster and up to 8 of those clusters share an L3 cache.

Similar concepts exist in device tree.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>