linux: don't look at packages if we're not the first PU of the core
"not first PU of core" always means "not first PU of die" on Linux.

Before Die support was added, this worked fine.
Now it works fine if the kernel exposes Dies.
But missing Die info in old kernels meant a package is parsed once
per core, because no PU was marked as first of a die.
Fix this.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>