lstopo/draw: cleanup line color management
Line color (for bridges) was always set to black in callers.
Also it was not correctly implemented in the xfig backend
(it was changing the fill color instead of the pen color).
Box border color was hardwired to black i the backends,
do the same for bridge lines.
It's unlikely that we'll ever need to change those.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>