windows: don't create groups for windows processor groups by default
As discussed in #497, Windows may now create processor groups
that strictly intersect with NUMAs or packages. This causes
a big warning when hwloc tries to insert the corresponding
group objects in the topology.

Fortunately, these groups aren't as useful as before in Windows 11
and WS2022. So disable them by default. We still have the Windows
specific API to query them, and group objs may be reenabled by setting
HWLOC_WINDOWS_PROCESSOR_GROUP_OBJS=1 in the environment.

Closes #497.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>