lstopo/windows: fix the initial window size by using the right fonts
The window initial dimensions are computed by drawing once
in an invisible window. We need to use the right font(s) to
get the correct dimensions that will match the final drawing
in the window.

Instead of creating a temporary font without specifying its
non-boldness (0 means FW_DONTCARE), define both normal and
bold fonts as usual during the initial step.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>