linux/cpukinds: ignore max frequencies is base frequency is known
Some Intel processors have Turbo Boost Max 3.0 which allows some cores to
have a slightly higher max turboboost frequency than others. For instance
i9-10920x gives 4.6GHz max to 8 cores, 4.7 to 2, and 4.8 to 2.
When the Linux backend sees this, it creates 3 cpukinds because max
frequencies are slightly different, while the base frequency is the same.

For now, we just ignore max frequencies is base frequencies are available
(which is the case on Intel CPUs with not-too-old Linux kernels).
We'll improve things later.

This disabling may be reverted with HWLOC_CPUKINDS_MAXFREQ=1.

maxfreq is forced back in tests/linux/fakeheterocpunuma
but disabled in tests/x86+linux/5intel64-hybrid-lakefield.

Refs #502.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>