linux/cpukinds: adjust max frequencies instead of ignoring them
If some cores have the same base frequency but different max freqs
(e.g. because of Intel Turbo Boost Max 3.0) adjust those max freqs
down to report the same for those cores.

By default, this is only enabled for differences < 10% but it may
be configured with HWLOC_CPUKINDS_MAXFREQ=adjust=X (X in percents)
or disabled (=1).
Ignoring (default behavior in 2.7) is still possible with =0.

Closes #502.

This is the real fix instead of the work-around in 2.7.0
in commit 46a7d94f3ab57ba56c9a8124dbf73ed147fbd155 .

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>